rename `styles.ts` into `pieStyles.ts`
