+++ packages/mermaid/src/styles.spec.ts
@@ -22,7 +22,7 @@ import er from './diagrams/er/styles.js';
 import error from './diagrams/error/styles.js';
 import git from './diagrams/git/styles.js';
 import gantt from './diagrams/gantt/styles.js';
-import pie from './diagrams/pie/styles.js';
+import pie from './diagrams/pie/pieStyles.js';
 import requirement from './diagrams/requirement/styles.js';
 import sequence from './diagrams/sequence/styles.js';
 import state from './diagrams/state/styles.js';
